Add files via upload
This is a Python project that was completed for the Principles of Programming 1 module. It is an object oriented program (due to it being coded using Java). It is a dungeon crawler that has an AI (bot) component that attempts to either catch the player or win the game itself. This was done with a priority level that was install in order to maximise the efficiency of the bot.